:truck: Rename Settings component, files and folders to SettingsScreen
