--- src/modules/Settings/components/index.js
@@ -1,1 +1,0 @@
-export { default as ItemTitle } from './ItemTitle';
